+++ storage/api/http.go
@@ -2,25 +2,13 @@ package api
 
 import (
 	"archive/zip"
-	"fmt"
-	"github.com/jinzhu/gorm"
 	"github.com/labstack/echo/v4"
 	"io"
-	"mime/multipart"
 	"net/http"
 	"os"
 	"strings"
 )
 
-
-type FileResponse struct {
-	ID       uint   `json:"ID"`
-	Filename string `json:"filename"`
-}
-
-type FilesResponse struct {
-	Files []FileResponse `json:"files"`
-}
 
 func ServeFile(c echo.Context) error {
 	db := GetDbConnection()
@@ -109,43 +97,4 @@ func StoreFile(c echo.Context) error {
 	}
 
 	return c.JSON(http.StatusCreated, response)
-}
-
-func storeFileToDB(path string, filename string, db *gorm.DB) (FileResponse, error) {
-	// Create a record in the DB as downloaded and send the kafka event.
-	var dbFile File
-	db.Where(&File{
-		Filename:   filename,
-		Path:       path,
-		Downloaded: true,
-	}).FirstOrCreate(&dbFile)
-
-	// Sending a kafka event.
-	//SendMessage(dbFile)
-
-	resp := FileResponse{ID: dbFile.ID, Filename: dbFile.Filename}
-
-	return resp, nil
-}
-
-func storeFileToDisk(fileFolder string, filename string, src multipart.File) (string, string, error) {
-	uniqueFilename, path := getUniqueNameAndPathFromFile(fileFolder, filename)
-
-	dst, err := os.Create(path)
-	if err != nil {
-		return "", uniqueFilename, err
-	}
-	defer dst.Close()
-
-	// Copying the file.
-	if _, err = io.Copy(dst, src); err != nil {
-		return "", uniqueFilename, err
-	}
-	return path, uniqueFilename, nil
-}
-
-func getUniqueNameAndPathFromFile(fileFolder string, filename string) (string, string) {
-	uniqueFilename := CreateUniqueFileName(filename)
-	path := fmt.Sprintf("%s/%s", fileFolder, uniqueFilename)
-	return uniqueFilename, path
 }
